backend/requirements.txt
Flask==3.0.0
Flask_JWT_Extended==4.5.3
Flask-Cors==4.0.0
PyMySQL==1.1.0
python-dotenv==1.0.0
